requirements.txt
flask
pandas
mysql.connector
json
datetime